Shorten verbose write_sac_ascii write commands.
If all numpy.savetxt values will have the same format, a single
formatter can be used, and the 'delimiter' keyword will be applied.  All
SAC header arrays *are* of the same format, per array, so verbose 'fmt='
flags were shortened, and the delimiter was specified to be empty. Also,
use savetxt for writing the string array, instead of a 'for' loop.