Fix byte order for header arrays vs data.
Header array byte order was previously initialized as little-endian or
native byte order.  This causes a mismatch when they are supposed to
correspond to a big-endian data vector, e.g. in SACTrace.from_obspy_trace.
Now, byteorder is a keyword in array initialization from scratch and
from a SAC header dictionary.