Change test_undefinedB behavior to retain reftime.
This test seems to make sure that unsetting tr.stats.sac.b doesn't
change the tr.stats.starttime when the trace is written and then re-read
in. It does this, however, by changing the SAC reference time ("nz"
header fields). You can see this in
self.assertEqual(sac_ref_time.timestamp, 269596800.0) and
self.assertEqual(sac_ref_time2.timestamp, 269596810.0).  Changing the
SAC reference time is not OK, because it invalidates all other relative
time headers in the file (i.e. a, o, and t0-t9, which are commonly used
to store phase picks and time windows). Also, if the file is iztype
"io", this is equivalent to changing an event origin time.  This test
preserves tr.stats.starttime when "b" is unset only by invalidating
other SAC headers.

When writing a SAC file from a Trace that was also read in from a SAC
file (has a tr.stats.sac), the old tr.stats.sac.b doesn't matter. One
only has to check that the old tr.stats.sac header has a valid reference
time, compare it to tr.stats.starttime, and then write a new "b" (and
"e") that respects this difference. No iztype checking or changes to the
SAC reference time is required, and any relative time headers therein
are preserved.

Here, I change this test such that it (1) makes sure that both
the tr.stats.starttime and the SAC reference time ("sac_ref_time") do
not change even if "b" is unset, and (2) allows "b" to become set when
writing from a Trace object.